% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cellType_list}
\alias{cellType_list}
\title{Dataset (list) containing markers for individual bone-related cell types.
List names indicate a given cell type. Each item contains associated markers.}
\format{
A list with 202 elements
}
\source{
\url{https://github.com/kkupkova/Mouse-bone-markers/blob/main/BONE_DATABASE_lists/bone_cellType.Rdata}
}
\usage{
cellType_list
}
\description{
Dataset (list) containing markers for individual bone-related cell types.
List names indicate a given cell type. Each item contains associated markers.
}
\keyword{datasets}
