% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cellTypeSource_list}
\alias{cellTypeSource_list}
\title{Dataset (list) containing markers for individual bone-related cell types
separated by source database. List names are in format source__cell type.
Each item contains associated markers.}
\format{
A list with 317 elements
}
\source{
\url{https://github.com/kkupkova/Mouse-bone-markers/blob/main/BONE_DATABASE_lists/bone_cellType_and_sourceDB.Rdata}
}
\usage{
cellTypeSource_list
}
\description{
Dataset (list) containing markers for individual bone-related cell types
separated by source database. List names are in format source__cell type.
Each item contains associated markers.
}
\keyword{datasets}
