% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{markerDatabase}
\alias{markerDatabase}
\title{Dataset containing bone-related markers.}
\format{
A data.frame with 133454 rows and 4 variables
\describe{
\item{tissue}{tissue of origin}
\item{cellType}{cell type or origin}
\item{gene}{marker}
\item{database}{source database from which makers were obtained}
}
}
\source{
\url{https://github.com/kkupkova/Mouse-bone-markers/blob/main/BONE_DATABASE.tsv}
}
\usage{
markerDatabase
}
\description{
Dataset containing bone-related markers.
}
\keyword{datasets}
