% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerGSEA.R
\name{markerGSEA}
\alias{markerGSEA}
\title{The function markerEnrich calculates enrichment of cluster-specific
variable markers within  bone cell type- or bone tissue-specific marker
database.}
\usage{
markerGSEA(
  varMarkers,
  systemLevel = "cellType",
  byDatabase = TRUE,
  minSize = 5,
  maxSize = 50000
)
}
\arguments{
\item{varMarkers}{A data.frame containing variable markers for each cluster
as reported by \code{FindAllMarkers} function from Seurat package.}

\item{systemLevel}{Optional string specifying on which systematic level to
perform enrichment, options: "cellType" or "tissue". Default: "cellType".}

\item{byDatabase}{Optional logical variable specifying if enrichment should
be performed separately for each database (i.e. source  from which markers
were obtained): option TRUE (default). Or if the database column in
\code{markerDatabase} should be ignored and enrichment is done on all
available markers for a given cellType or tissue: option FALSE.}

\item{minSize}{Number specifying a minimal number of markers within a
marker set to be considered (defaults to 5).}

\item{maxSize}{Number specifying a maximal number of markers within a
marker set to be considered (defaults to 50000).}
}
\value{
data.frame
}
\description{
The function markerEnrich calculates enrichment of cluster-specific
variable markers within  bone cell type- or bone tissue-specific marker
database.
}
\examples{
varMarkerFile = system.file("extdata", "cluster_variable_features.csv",
                            package = "BoneCellType")
varMarkers = read.csv(varMarkerFile)
GSEAres = markerGSEA(varMarkers)
}
