% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerEnrichment.R
\name{plotMarkerEnrich}
\alias{plotMarkerEnrich}
\title{The function plotMarkerEnrich plots results from \code{markerEnrich} function}
\usage{
plotMarkerEnrich(enrichRes, topN = Inf)
}
\arguments{
\item{enrichRes}{An output from \code{markerEnrich} function in form of a
data.frame.}

\item{topN}{And optional parameter specifying the number of the most
significant results (based on p-value) to be plotted per cluster.}
}
\value{
A ggplot object
}
\description{
The function plotMarkerEnrich plots results from \code{markerEnrich} function
}
\examples{
varMarkerFile = system.file("extdata", "cluster_variable_features.csv",
                            package = "BoneCellType")
varMarkers = read.csv(varMarkerFile)
enrichRes = markerEnrich(varMarkers)
plotMarkerEnrich(enrichRes)
}
