% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerGSEA.R
\name{plotMarkerGSEA}
\alias{plotMarkerGSEA}
\title{The function plotMarkerGSEA plots results from \code{markerGSEA} function}
\usage{
plotMarkerGSEA(enrichRes, topN = Inf, plotNES = FALSE)
}
\arguments{
\item{enrichRes}{An output from \code{markerGSEA} function in form of a
data.frame.}

\item{topN}{An optional parameter specifying the number of the most
significant results (based on p-value and NES) to be plotted per cluster.}

\item{plotNES}{and optional logical variable if bar height should be NES
instead of a -log10(padj), since padj can be really low}
}
\value{
A ggplot object
}
\description{
The function plotMarkerGSEA plots results from \code{markerGSEA} function
}
\examples{
varMarkerFile = system.file("extdata", "cluster_variable_features.csv",
                            package = "BoneCellType")
varMarkers = read.csv(varMarkerFile)
enrichRes = markerGSEA(varMarkers)
plotMarkerGSEA(enrichRes)
}
