% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tissue_list}
\alias{tissue_list}
\title{Dataset (list) containing markers for individual bone-related tissues.
List names indicate a given tissue. Each item contains associated markers.}
\format{
A list with 4 elements
}
\source{
\url{https://github.com/kkupkova/Mouse-bone-markers/blob/main/BONE_DATABASE_lists/bone_tissue.Rdata}
}
\usage{
tissue_list
}
\description{
Dataset (list) containing markers for individual bone-related tissues.
List names indicate a given tissue. Each item contains associated markers.
}
\keyword{datasets}
