% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tissueSource_list}
\alias{tissueSource_list}
\title{Dataset (list) containing markers for individual bone-related tissues
separated by source database. List names are in format source__tissue.
Each item contains associated markers.}
\format{
A list with 11 elements
}
\source{
\url{https://github.com/kkupkova/Mouse-bone-markers/blob/main/BONE_DATABASE_lists/bone_tissue_and_sourceDB.Rdata}
}
\usage{
tissueSource_list
}
\description{
Dataset (list) containing markers for individual bone-related tissues
separated by source database. List names are in format source__tissue.
Each item contains associated markers.
}
\keyword{datasets}
